# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kizuna Rail API",
    "version": "1.0.0",
    "description": "API for the Kizuna Rail website."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "securitySchemes": {
      "SessionCookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Express session cookie set after login."
      }
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log in a user",
        "description": "Authenticates a guest. Requests with an active session are rejected; log out before switching accounts.",
        "security": [
          {},
          {
            "SessionCookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "email": {
                        "type": "string",
                        "format": "email",
                        "example": "hector@example.com"
                      },
                      "password": {
                        "type": "string",
                        "format": "password",
                        "example": "mysecurepassword"
                      }
                    },
                    "required": [
                      "email",
                      "password"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "username": {
                        "type": "string",
                        "example": "hector"
                      },
                      "password": {
                        "type": "string",
                        "format": "password",
                        "example": "mysecurepassword"
                      }
                    },
                    "required": [
                      "username",
                      "password"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully."
          },
          "400": {
            "description": "Invalid input data."
          },
          "401": {
            "description": "Invalid email or password."
          },
          "409": {
            "description": "A user is already logged in in this session."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log out a user",
        "description": "Logs out the currently authenticated user.",
        "security": [
          {
            "SessionCookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User logged out successfully."
          },
          "400": {
            "description": "No user is currently logged in."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new user",
        "description": "Guests and admins may create a customer account. Signed-in customers are not allowed to register another account.",
        "security": [
          {},
          {
            "SessionCookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Hector"
                  },
                  "username": {
                    "type": "string",
                    "example": "hector"
                  },
                  "email": {
                    "type": "string",
                    "example": "hector@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "mysecurepassword"
                  }
                },
                "required": [
                  "name",
                  "username",
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully."
          },
          "400": {
            "description": "Invalid input data."
          },
          "403": {
            "description": "Signed-in customer cannot register another account."
          },
          "409": {
            "description": "Email is already in use."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List all users",
        "description": "Admin only. Password hashes are never returned.",
        "security": [
          {
            "SessionCookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Users returned successfully."
          },
          "401": {
            "description": "Missing or invalid session."
          },
          "403": {
            "description": "Admin role required."
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Register a user",
        "description": "Guests and admins may create customer accounts. Signed-in customers are forbidden.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "username",
                  "email",
                  "password"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Hector"
                  },
                  "username": {
                    "type": "string",
                    "example": "hector"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "hector@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "mysecurepassword"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully."
          },
          "400": {
            "description": "Required registration data is missing or invalid."
          },
          "403": {
            "description": "Signed-in customers cannot create accounts."
          },
          "409": {
            "description": "Email is already in use."
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get a user by ID",
        "description": "The user may retrieve their own record; admins may retrieve any user.",
        "security": [
          {
            "SessionCookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB user ID.",
            "schema": {
              "type": "string"
            },
            "example": "6ab9358a87b7afeb8a3794b1"
          }
        ],
        "responses": {
          "200": {
            "description": "User returned successfully."
          },
          "400": {
            "description": "Invalid user ID."
          },
          "401": {
            "description": "Missing or invalid session."
          },
          "403": {
            "description": "The caller is neither this user nor an admin."
          },
          "404": {
            "description": "User not found."
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update a user",
        "description": "Users may update their own name and email. Admins may update any user and may also set role.",
        "security": [
          {
            "SessionCookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "6ab9358a87b7afeb8a3794b1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Hector"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "hector@example.com"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "customer",
                      "admin"
                    ],
                    "description": "Admin only.",
                    "example": "customer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully."
          },
          "400": {
            "description": "Invalid ID or user information."
          },
          "401": {
            "description": "Missing or invalid session."
          },
          "403": {
            "description": "Caller is not authorized or cannot change roles."
          },
          "404": {
            "description": "User not found."
          },
          "409": {
            "description": "Email is already in use."
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete a user",
        "description": "Users may delete their own account; admins may delete any account. Associated bookings are deleted too.",
        "security": [
          {
            "SessionCookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "6ab9358a87b7afeb8a3794b1"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully."
          },
          "400": {
            "description": "Invalid user ID."
          },
          "401": {
            "description": "Missing or invalid session."
          },
          "403": {
            "description": "Caller is neither this user nor an admin."
          },
          "404": {
            "description": "User not found."
          }
        }
      }
    },
    "/api/roles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "List roles",
        "description": "Admin only.",
        "security": [
          {
            "SessionCookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Roles returned successfully."
          },
          "401": {
            "description": "Missing or invalid session."
          },
          "403": {
            "description": "Admin role required."
          }
        }
      }
    },
    "/api/roles/user/{userId}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Get a role by user ID",
        "description": "Admin only.",
        "security": [
          {
            "SessionCookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "64b8f0c2e1b2a3d4f5678901"
          }
        ],
        "responses": {
          "200": {
            "description": "Role returned successfully."
          },
          "400": {
            "description": "Invalid user ID."
          },
          "401": {
            "description": "Missing or invalid session."
          },
          "403": {
            "description": "Admin role required."
          },
          "404": {
            "description": "User or role not found."
          }
        }
      }
    },
    "/api/trains": {
      "get": {
        "tags": [
          "Trains"
        ],
        "summary": "Get all trains",
        "description": "Returns every train in the trains collection",
        "responses": {
          "200": {
            "description": "Trains retrieved successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/schedules": {
      "get": {
        "tags": [
          "Schedules"
        ],
        "summary": "Get all schedules",
        "description": "Returns every schedule in the schedule collection",
        "responses": {
          "200": {
            "description": "Schedules retrieved successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/stations": {
      "get": {
        "tags": [
          "Stations"
        ],
        "summary": "Get all stations",
        "description": "Returns every station in the stations collection",
        "responses": {
          "200": {
            "description": "Stations retrieved successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}